To make it compatible with StackExchange.Redis .NET
The StackExchange.Redis .NET lib is doing a PING, EXISTS and GET
command just after established the tcp connection. We include the
support of this commands and also the option to ignore commands defined
in the configuration (ignoreCommands):

[[relayer]]
protocol = "firehose"
mode = "smart"
listen = "unix:/tmp/firehose.sock"
maxConnections = 2
maxRecords = 100
streamName = "testing"
region = "eu-west-1"
ignoreCommands = "SET GET"

The commands included in the list of ignoreCommands will response as "1"
(true) and stop the process. It will not send anything to firehose.

We are doing this because StackExchange.Redis does a SET before each op
to redis.